requirements.txt
python-dotenv==1.0.1
openai==1.25.0
streamlit==1.33.0
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
PyPDF2==3.0.1
faiss-cpu==1.8.0
altair==4
tiktoken==0.5.2
chromadb==0.3.29
beautifulsoup4==4.12.3
watchdog==4.0.0


#qdrant_client